requirements.txt
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
python-dotenv==1.1.1
python-telegram-bot==20.7
requests==2.32.5
sniffio==1.3.1
urllib3==2.5.0
